Conditional rendering of libraryHubPage
If user is a Teacher will display BookAddPage else LibraryChoicePage.
BookAddPage is currently blank
