+++ library-app/client/src/pages/BookAddPage.jsx
@@ -1,0 +1,10 @@
+import React from "react"
+import {useMutation, useQuery} from '@apollo/client';
+import { QUERY_LIBRARY } from "../utils/queries";
+import { Link } from "react-router-dom";
+
+const BookAddPage = () => {
+
+}
+
+export default BookAddPage
